Updated releng/autorelease Project: snmp4sdn  0af455339817464aadbc0830d267f3b9577bca3c
Set root pom.xml <name> for Sonar.

As mentioned on the mailing list Sonar uses the <name> field of the
pom.xml that is passed to the mvn command as the name of the project in
Sonar. In most cases this is the root pom.xml file in a project. This
patch sets the name to the project shortname.

https://lists.opendaylight.org/pipermail/discuss/2014-November/004024.html

Change-Id: I93b6db7bfc5d578231d9926cb19894151b4ce394
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>