Updated releng/autorelease Project: snbi  5a6236c374bb6cfa34c90fbdc3426c8a72f525d4
Set root pom.xml <name> for Sonar

As mentioned on the mailing list Sonar uses the <name> field of the
pom.xml that is passed to the mvn command as the name of the project in
Sonar. In most cases this is the root pom.xml file in a project. This
patch sets the name to the project shortname.

https://lists.opendaylight.org/pipermail/discuss/2014-November/004024.html

Change-Id: I494ef1dc5eb838cb2e4308765e7409be8b790ea8
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>