Updated releng/autorelease Project: dlux  05bd69a36fe2ab326006a010073296b188b46034
Fix formatting for property

Change-Id: I8802b31eda031ab46ae4f37c568309b11fc0be42
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>

Set root pom.xml <name> for Sonar

As mentioned on the mailing list Sonar uses the <name> field of the
pom.xml that is passed to the mvn command as the name of the project in
Sonar. In most cases this is the root pom.xml file in a project. This
patch sets the name to the project shortname.

https://lists.opendaylight.org/pipermail/discuss/2014-November/004024.html

Change-Id: I1a1e76da6dd596a0b5e916f7ddcda72d3720b58d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>