Updated releng/autorelease Project: integration  3c72f24b1e26c642028c3aa2ad01dbd96037df67
Update checkstyle version to 2.13

Change-Id: I01fe69474fc3b07972b5080e00a6284b10215fd1
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>

Use maven-enforcer-plugin to enforce Maven version

This patch removes the unnecessary declaration of <prerequisites>
section.

As discovered in the mailing list thread linked below the <prerequisites>
section to set a minimal Maven version is deprecated and actually does
not enforce the minimal version. If we need to enforce a minimal Maven
version we need to use the maven-enforcer-plugin instead.

https://lists.opendaylight.org/pipermail/discuss/2015-January/004309.html

Change-Id: I5e5ce1af34451b0bcda51ee10ee0c181bd85393b
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>