Updated releng/autorelease Project: l2switch  6e0d3e9a42f44916c465efaff7b24d1ba4dd2f1c
Set root pom.xml <name> for Sonar

As mentioned on the mailing list Sonar uses the <name> field of the
pom.xml that is passed to the mvn command as the name of the project in
Sonar. In most cases this is the root pom.xml file in a project. This
patch sets the name to the project shortname.

https://lists.opendaylight.org/pipermail/discuss/2014-November/004024.html

Change-Id: I48552988c06e856136a4bdfc068f40e7e2f352a5
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>