Updated releng/autorelease Project: bgpcep  b4127185fc06f6015dd86a456b405858185468b8
Optimize AsPath lists

Tracing on 500K routes has shown we waste about 35MB of memory in sparse
arrays. It turns out that we fail to size ArrayLists properly when
allocating AsPaths: in case of attribute, we can trimToSize(), in case
of individual segments we know the right size beforehand.

Change-Id: Ie25610b9f04e57c6659b69b65f20299c5cb10881
Signed-off-by: Robert Varga <rovarga@cisco.com>

Fix a potential deadlock in BGPSession

In a two-session scenario, it is possible to arrive at a ABBA deadlock,
where we acquire locks:

Thread 1      Thread 2
--------      --------
Session A     Session B

RIB
              RIB <-- blocks waiting for Thread 1

Session B <-- blocks waiting for Thread 2

The lock is being held only for scheduling a task on Thread 2, which a
thread-safe operation, so we can simply remove the need to take the lock
in Thread 1.

Change-Id: I4e55d3228a48790b2f19eef12ce1b54fd57955a0
Signed-off-by: Robert Varga <rovarga@cisco.com>