Updated releng/autorelease Project: controller  4242eef391e1e35f45eaac6e95ccd7766f9c5c6d
Lower Shard.DEFAULT_NAME visibility

Add an annotation which clarifies that the DEAFULT_NAME is only ever
used in tests.

Change-Id: I786a2e747e4f52ac4577a94fd268e13a191e07e9
Signed-off-by: Robert Varga <rovarga@cisco.com>

Use ImmutableSet in ShardManager

Instead of mutating the same set, use a space-optimized immutable view
and update that.

Change-Id: Iae0615580f93bb8e215d929e72abeb17f8b3eed0
Signed-off-by: Robert Varga <rovarga@cisco.com>

Clarify DistributedDataStoreFactory

DistributedDataStoreFactory used an atomic reference to manage a lazy
initialization of a singleton, which could fail to initialize. Use an
explicit volatile and lower synchronization requirements when the
singleton has already been initialized.

Change-Id: I10c38ca2441c6e7d8251560f9ea1218211490fc9
Signed-off-by: Robert Varga <rovarga@cisco.com>