Updated releng/autorelease Project: controller  989c3723e634bf4a0fb671246c31eadf255c462c
Hide internals of FollowerLogInformationImpl

The constructor should not take AtomicLong instances, as that is an
implementation-specific detail. Notably the constructor assumes
ownership of the objects and no caller uses them in a shared way.

Change-Id: I587fb2033f9b513df565471f037e20e0acc481c7
Signed-off-by: Robert Varga <rovarga@cisco.com>

Do not leak AtomicLong from FollowerLogInformation

Leaking this implementation details allows callers to interact with
state outside of the official contract. Disallow that.

Change-Id: I880310f2a9692fd11ba2cd33830501d19d0a1b65
Signed-off-by: Robert Varga <rovarga@cisco.com>

Speedup AbstractLeader.printFollowerStates()

Instead of using string concat to append to a StringBuilder and then
concat the result of the builder, perform proper construction and
instantiate only a single string.

Change-Id: I05f0bd9404a3d6470a0b9d0b1ea71595dcba5cc0
Signed-off-by: Robert Varga <rovarga@cisco.com>

Hide AbstractLeader maps

Collections should never be leaked, as they introduce the ability to
make unsynchronized changes. This patch hides the map itself, allowing
lookups.

Once we have done that, we have control over the interactions and thus
can see that the map is instantiation-constant. Use an ImmutableMap to
maintain the entries as it provides superior compactness and lookup
speeds.

Change-Id: I552372dbabe454111909bb8251637e9f3c0e0db7
Signed-off-by: Robert Varga <rovarga@cisco.com>

Remove followers as an explicit field

The field corresponds to followerToLog.keySet(), which is immutable, so
there is no point in retaining it in a field. It allows us to realize
that we can use an entrySet() and prevent lookups to get corresponding
information.

Change-Id: I9f43732813b61d19306ef5954a97c9c8cd64ccfb
Signed-off-by: Robert Varga <rovarga@cisco.com>

Improve RpcInvocationStrategy performance

As it turns out, the transformation function can easily be shared,
reducing the amount of objects we create.

Change-Id: I76c86e2f0c63e87c87677ac739f980a5e7929d3f
Signed-off-by: Robert Varga <rovarga@cisco.com>