Updated releng/autorelease Project: controller  3f153e5fa694fe4147e72e615edbb5c263e5a394
Use an ImmutableList instance for getFrom()

Since we are creating a view, we do not need to allocate a larger list
-- just use a copy constructor, which will create an
appropriately-sized copy.

Change-Id: I6a4b90f638b3f8511b5ec641ab5f9b1f14632407
Signed-off-by: Robert Varga <rovarga@cisco.com>

Improve AbstractLeader tracker removal

The remove operation performs a linear lookup by index and perform a
second iteration to find the object again. Instead of that, instantiate
an iterator and use its remove() method to elide the second lookup.

Change-Id: Ida2a5c327115d2da787e397311c0a0c5a46a0d43
Signed-off-by: Robert Varga <rovarga@cisco.com>

use AtomicLongfieldUpdater in FollowerLogInformationImpl

Instead of keeping two AtomicLong instances, we instantiate two shared
updater classes and keep the values in volatile fields. This lowers the
memory consumption.

Change-Id: I9e6ce426500ee61b8c02d75cb0edd361e996a443
Signed-off-by: Robert Varga <rovarga@cisco.com>

Change trackerList to a LinkedList

The trackers are being added to the end and usually remove in a linear
fashion. An ArrayList's remove() method makes sure the backing array is
packed -- which means we end up copying the array over and over again.

Changing the trackerList to a LinkedList removes the need for copying at
the expense of memory footprint and cache locality.

Change-Id: I3e40487798eb64713ed4526f53a0a8c385a9a778
Signed-off-by: Robert Varga <rovarga@cisco.com>

Hide AbstractReplicatedLogImpl index fields

The fields have proper getter/setter methods, there is no need to expose
them anywhere.

Change-Id: I7971753310dfe8c8654caeef04cfc38dfddeef86
Signed-off-by: Robert Varga <rovarga@cisco.com>