Updated releng/autorelease Project: controller  6bcc799fb62c2e4572f11134a601e7ea2a4651cf
Speedup ModuleShardStrategy a bit

Using isEmpty() is a bit faster, as it allows us to side-step a
subsequent comparison.

Change-Id: If2470ac293f0678e7bf158fccc00f263ed73a207
Signed-off-by: Robert Varga <rovarga@cisco.com>

Optimize getAllShardNames()

Instead of iterating through all shards, cache the result of
getAllShardNames().

Change-Id: Ia36e2a7cbae39e7e6862e57d93532b7b59c7863b
Signed-off-by: Robert Varga <rovarga@cisco.com>

Optimize getModuleNameToShardStrategyMap()

This map is based on constant data and can be easily pre-computed. Do
precisely that.

Change-Id: Ic3851034ca3ccd752c3f4afc0d1d3c77a9df9e65
Signed-off-by: Robert Varga <rovarga@cisco.com>

Optimize getShardNamesFromModuleName()

Instead of creating the list each and everytime, take advantage of the
fact the inputs are constants -- and pre-instantiate a map.

Change-Id: If754942561a186fa1ca7806fe86834a3530a97ca
Signed-off-by: Robert Varga <rovarga@cisco.com>

Use a singleton instance for DefaultShardStrategy

Instead of instantiating the DefaultShardStrategy everywhere it is used,
share a singleton instance.

Change-Id: I24ebd4b1d713a09b4fc4161e9c8c69a088af5ac6
Signed-off-by: Robert Varga <rovarga@cisco.com>

Make sure QNameFactory cache is constant

The cache is constant, there is no reason it should be non-final.

Change-Id: I663352473e6a9e6021c3cee56ddf0336e5f8e2ac
Signed-off-by: Robert Varga <rovarga@cisco.com>

Optimize getModuleNameFromNameSpace

Instead of looking up in the global array which is constant,
pre-construct a constant lookup map.

Change-Id: I85d54332b0a3b717d33da9758eaafbce5bd5595d
Signed-off-by: Robert Varga <rovarga@cisco.com>