Updated releng/autorelease Project: ovsdb  03a9ed9eaa037e52585b59ec84fde1156d9539c3
Re-enable southbound build.

After some investigation, this turns out to be a deeply adverse interaction
between pax-exam 4.4.0 (and 4.0.0-4.4.0) and mirrors in the infra.

Since I believe we need to go to at least pax-exam 4.0.0 to get to karaf
3.0.2 to get to Java 8 (or so I recall from talking to Robert), the mirror
thing in infra has been fixed (nobodies preferred fix).

This patch should now verify, and the build should be restored.

Change-Id: I993eae3cfe999faa5a6262e61144bf60e5be261c
Signed-off-by: Ed Warnicke <eaw@cisco.com>