Updated releng/autorelease Project: controller  fdc585863663ca5a4baf83ef146390e3c845a567
Optimized AppendEntries sending to follower by adding AppendEntry call at end of reply, if needed.

This will help to reduce the lag of a follower and leader does not have to wait for heartbeat time to
send a message to follower. This way, we can tune heartbeat interval in better manner.

Change-Id: I0259ea3691a3ac1245e36afe15b06ed44f377466
Signed-off-by: Harman Singh <harmasin@cisco.com>