Updated releng/autorelease Project: controller  e3918142bf119759d092af5c7fa72807b3a467e0
Bug-2692:Avoid fake snapshot during initiate snapshot

When an initiate snapshot is received , we should be setting the snapshot captured flag, in order to avoid a fake snapshot at the same time.

Also a check should be made to avoid another initiate snapshot to occur, in case the capture snapshot takes time.

Change-Id: I6e29c97b8cc71c101504cb3588897a896e107e93
Signed-off-by: Kamal Rameshan <kramesha@cisco.com>