Updated releng/autorelease Project: ovsdb  43c328b63ae1f680a9e9921dc677f2d35c808798
OVSDB MD-SAL SB - remove extra disconnected call

When deleting an active ODL connection to an OVS node
via restconf, the disconnect() call in
OvsdbConnectionManager.java called the disconnected()
method which resulted in two transactions to remove the
operational MD-SAL entry for the node.  Which resulted
in WARNings in the log.

Change-Id: I305be04a9217c03d646b103294e7f0d7eabf7df8
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>