Updated releng/autorelease Project: controller  05f8f86227bfd107e536684f9c593f5aeacb2811
Refactor DataStoreContext

It was getting difficult to maintain because we had the same field defined
in the Builder and the DatastoreContext and we were passing the fields along
using the DatastoreContext constructor. This meant that every new setting that
we added required multiple changes to be made to DatastoreContext.

Refactored for the Builder to create the DatastoreContext upfront and directly set
it's fields.

Change-Id: I8060c4ea3dfd02a5cac20e28fb5be2f26b1e1dbd
Signed-off-by: Moiz Raja <moraja@cisco.com>

Store the datastore type in dataStoreContext

This way we do not need to pass it to DistributedDataStoreFactory,
DistributedDataStore and ShardManager as it is already present in
dataStoreContext

Change-Id: Id9aa1541728f3632a46d67f3bdbacb87fd39783f
Signed-off-by: Moiz Raja <moraja@cisco.com>

BUG 2676 : Apply backpressure when creating transactions

This patch uses a rate limiter to limit the number of write transactions
that can be created on the datastore at any given point of time (within a
second).

We also have added a metric that is trackable using JMX which will show
exactly how much time we take in committing a transaction.

Could this rate limiter have been on the DataBroker? Yes it could have been
but I think we need to go granular on this one. Ideally the ratelimiter would
be for every shard transaction but it's not trivial to do in a way that would
actually block the consumer for now - this is because creating a transaction on
Shard itself is an async operation.

Change-Id: I3d8e3baeb892ba494c0ab0d13a0d6226c1516511
Signed-off-by: Moiz Raja <moraja@cisco.com>