Updated releng/autorelease Project: ovsdb  c05f08b4cdea4ea2271da0c3d0545640f65b3d16
Bug 2707: Pipeline flows are not programmed because of failed onNodeUpdated call

Cause: we used the onNodeUdpated() OpendaylightInventoryListener interface to receive updates when openflow nodes were updated. If we thought it was the first time and the openflow node was just added we would then program the pipeline flows. ovsdb checked the FlowCapableNode in the datastore for the isPreset flag. It looks like there might have been some race condition where another writer like InventoryManager was updated the FlowCapablNode such that when ovsdb would do the above check isPresent would return true and ignore the update - and not program the pipeline flows.

Solution: replace the deprecated call with an mdsal flowCapablenode DataChangeListener. This ensures ovsdb will get the node when created.

Also added a couple log.info's to track when the nodes are added to help with future debugging.

Change-Id: Ib6867927cb1ce61670d7566ea40b88966243bdca
Signed-off-by: Sam Hague <shague@redhat.com>