Updated releng/autorelease Project: controller  97307fbbbee1f9bdb1409ac962386779dc4e93bf
Make ReadData exception logging less noisy

There is some component (openflow) which is reading data on a ClosedTransaction.
This is causing logging on the datastore side. The component itself seems to handle
the error reported by the datastore quietly, so switching the logging the error to
debug and logging full stack trace at trace level.

Change-Id: I3c118765dca7f4a50d26c8dd50eee96c65de1871
Signed-off-by: Moiz Raja <moraja@cisco.com>