Updated releng/autorelease Project: groupbasedpolicy  fa233c807e53a3775017d7ce6691032bf0d8175c
Fix build after openflowplugin merge.

This fixes a build breakage caused by a merge in the openflowplugin project
which enabled flow programming models/features.

This patch also re-enables the single-feature-test that
was broken when moving this from yangtools to odlparent.

Lastly, this patch removes the OpenContrail feature from
automatic startup (we're migrating to per-renderer features
in a subsequent patch).

Change-Id: I09382b2cef32efbd542d45c070280e1963858c80
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>