Updated releng/autorelease Project: controller  6051cbabfe99c494b4f3b75aca516fb6f27338ad
Fix a race in PingPong transaction scheduling

We have failed to re-check inflighTx after taking the lock, which meant
that completion and readiness may have raced resulting in warnings about
transaction being submitted while there are others in-flight.

Fix also the shutdown case, which could trigger an unexpected warnings
if there is an inflight transaction.

Change-Id: Ie3a3845335754860a831070d5a73dbc9a4d67d3c
Signed-off-by: Robert Varga <rovarga@cisco.com>