Updated releng/autorelease Project: controller  7d03757a62c1909ee74b42e6acf30d3d7b961bfa
BUG 2718 : Create a diagnostic utility to track append entries replies

This patch introduces a utility to keep track of all messages that were
processed between the time a certain message is received. It is used for
tracking why AppendEntriesReply do not come in a timely manner. This is
important to know because if a leader will switch to IsolatedLeader if
these messages do not arrive within the ElectionTimeout interval causing
disruption in service.

Change-Id: I13bab798aad8a225d22d2ebd3775cbf0bf1d1592
Signed-off-by: Moiz Raja <moraja@cisco.com>