Updated releng/autorelease Project: yangtools  1ef225ddeed34c62e37e596aead4921c4c26af25
YANG JSON codec can be used inside other JSON stream.

This is useful for emitting root element, which is
not present in YANG schema.

This also allows to introduce special support of
emitting get-config and get RPC bodies, which
are using data tree schema for anyxml, and
emitted stream is still considered valid YANG JSON
according to specification.

Change-Id: Ie01afc93834bf4e9ee4981a71413ae477dee0766
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>