Updated releng/autorelease Project: controller  6345946ab68108ac70de6f0c472b520802656abe
Bug 2763: Fixed JSON serialization of root of mount point

Upgrade of yang-data-codec-gson library to use GSON
for streaming and serialization of JSON broke
one special-case which was reading of mount point
root.

This patch migrates NormalizedNodeJsonCodec to
use proper APIs, which allows more explicit control
of how stream is processed.

Change-Id: I0f57e8ef99114a08b1bceb51c47efbcb0546af73
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>